Make Scheduler's default args public
